Move to registering Checks rather than hardcoding
Rather than hardcode the list of checks within the linter we move to
registering them, this will allow us to skip checks if we so wish and
also better manage the dependencies of the Lint class.

Not sure if the top level is the best place to register these but it'll
do for now.