Sub liquid back into HTMLCheck
We have to remove liquid tags from the document for us to be able to run
the Nokogiri parsing on it, originally we did this by subbing in spaces
in the place of liquid tags, this led to some odd error messages with
big gaps where there would've been liquid tags.

Here we move to an approach similar to that used by Shopify in their
theme check tool to replace the tag with a keyed placeholder and then
sub those back in based on the key when we present the error back to the
user.

We use base 36 to make sure that we dont overflow the lower limit of 4
characters ( e.g. {{}} or {%%} ) when we have too many tags on a page,
for base 10 this would be 1000 which seems doable, 36 gives us way more
  at 46656 tags in a document.