Allow layout to have accordion_toggle element
The `accordion_toggle` element will act similarly to the `accordion`
element, where it shows/hides a list of attributes. However, it will be
controlled by a boolean variable.

For example, if the use the user toggles "Show CTA", then we want to
reveal the variables associated with the CTA, such as "CTA text"
and "CTA link".