Update LayoutSchema so it expects a hash of attributes
Original schemas define `attributes` as a hash and we use
the ExpandAttributes service in various parts of this gem in order
to "expand" this hash into an array.

Some classes expect the schema to already have the attributes "expanded"
and some don't. In order to make it consistent, the aim to have all classes
expect the original schema, where the attributes have not been expanded.

It is then up to the class itself to use the ExpandAttributes service if it
needs to in order to have an array of attributes.