Refactor when attributes are expanded in MenuSchema validator
In the same way we did for BlockSchema, this commit changes when
we "expand" the attributes. We will now leave the schema in its original
format and only use the ExpandAttributes service within a private method
when we need to have attributes as an array in order to validate them.