Bump version to v2.0.0
This version allows menu and footer schemas to also
have a `layout` key, in the same way as blocks.

Bumping the major version here because now the MenuSchema and
FooterSchema validators expect the `attributes` in the schema to be
in a hash format, where previously they expected an array format.

This change is not backwards compatible.