Install dartsass-rails as a dependency
Despite its name, it doesn't actually pull in all of rails. Its just a
wrapper around the dartsass binary with some useful rake tasks. In this
project we only care about the binary, but using this as a dependency
means we only pull in one dartsass binary when this is used in a rails
application.