Lint before syncing
I'm not proud of the code here, but cli ui sucks and makes outputting
stuff a pain in the ass.
What this does is effectively just run the linter on the changed file
before we sync it, most of the changes are just making the canvas gem
not output so that we can manage in the context of the new spin group or
something?