Ensure canvas lint fails if an error is raised
At the moment, canvas lint will fail a check if the check explicitly
adds an Offense. However, if an unexpected error occurs, this gets
swallowed by the CLI::UI::SpinGroup and it does not consider the check
as failed.

This causes a problem on the GitHub check, in that the lint check is
green on the GitHub PR even though an error has occurred.

This commit solves the problem by catching any unexpected errors,
printing these out and returning `exit 1` in this situation.