Refactor Canvas::Lint
This commit refactors the Canvas::Lint class so that there is a single
@failed_checks variable (rather than a separate one for
@errored_checks as well).

We need to run the check outside of the `spin_group.add` block
because otherwise the error handling doesn't play nicely with
the CLI::UI::SpinGroup class.

Also the DummyClass needs to define a `run` method in the spec,
otherwise a `NotImplementedError` is raised.