Enforce lowercase keys for custom types
We store custom types in the database with a requirement to have keys
unique within a theme.

This validation is case sensitive; currently, it's possible to create
multiple version of the same type with names like `type` and `Type`.

Changing this validation to case insensitive in the persistence layer
could lead into themes failing to load custom types in the future. (Now
it's guaranteed to be unique as the first time load from the theme
creates the first custom types.)

To ensure integrity across all layers, we want to enforce lowercase keys
at the point of custom type validation.

Co-authored-by: Lizzy Ballantine <lizzy.kayballantine@gmail.com>