Add experience_slot as a primitive type
This commit allows the `experience_slot` variable to be used
in schemas. It allows a default value of "next_upcoming" e.g.

---
attributes:
  my_slot:
    type: experience_slot
    default: next_upcoming
---

Also, I'm updating the wording in the experience_date specs, as it
was pointed out in review that the same ones for experience_slot were
not as clear as they could be.