Upgrade ruby version for Github Action
Cause:

With ruby 3.1 there was outdated rubygems included which with never
version of psych gem version was failing to install one of the gems.

https://github.com/easolhq/easol-canvas/actions/runs/11702680056
https://github.com/openstreetmap/openstreetmap-website/issues/4364

Solution:

One of the way to fix that problem was to update in Github action
rubygems, but I opted in for upgrading Ruby version itself which already
comes with proper version of rubygems and is already used in main Easol
app.

Also upgraded Rubocop's ruby target version to match Ruby version.