Allow html elements to be set by liquid variables
This is something our theme devs want to do. Previously, we replaced
liquid variables with empty spaces, which is fine for most uses but
when it's used directly in the html tag this results in:
`ERROR: Invalid first character of tag name ' '.`

By changing this to `x` we can keep the same number of characters but
allow for liquid variables to be used in the tag.